Week2/LongProject/backend/requirements.txt
fastapi
uvicorn
python-multipart
pinecone
langchain
langchain-community
langchain-pinecone
langchain-huggingface
groq
ddgs
google-generativeai
sentence-transformers
rank_bm25
pillow
pyyaml
opentelemetry-sdk
opentelemetry-exporter-otlp
langsmith
requests
python-dotenv
arxiv
googlesearch-python
networkx
pydantic
pytesseract
PyPDF2